ML homework1

+++ chap02/homework/tm.pyx
@@ -48,13 +48,13 @@ cdef void target_mean_v3_impl(double[:] result , double[:] y , double[:] x , con
 
 cpdef target_mean_v4(data, y_name, x_name):
     cdef long nrow = data.shape[0]
-    cdef np.ndarray[double] result = np.asfortranarray(np.zeros(nrow), dtype=np.float64)
+    cdef np.ndarray[float] result = np.asfortranarray(np.zeros(nrow), dtype=np.float32)
     cdef np.ndarray[int] y = np.asfortranarray(data[y_name], dtype=np.int32)
     cdef np.ndarray[int] x = np.asfortranarray(data[x_name], dtype=np.int32)
     target_mean_v4_impl(result , y, x , nrow)
     return result
 
-cdef void target_mean_v4_impl(double[:] result , int[:] y , int[:] x , const long nrow):
+cdef void target_mean_v4_impl(float[:] result , int[:] y , int[:] x , const long nrow):
     cdef dict value_dict = dict()
     cdef dict count_dict = dict()
